requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aistudio-sdk==0.3.8
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
bce-python-sdk==0.9.52
beautifulsoup4==4.14.2
cachetools==6.2.2
certifi==2025.10.5
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
colorlog==6.10.1
cssselect==1.3.0
cssutils==2.11.1
dataclasses-json==0.6.7
distro==1.9.0
einops==0.8.1
et_xmlfile==2.0.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
ftfy==6.3.1
future==1.0.0
fuzzywuzzy==0.18.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.2.0
html2image==2.0.7
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.1.4
idna==3.11
imagesize==1.4.1
imgkit==1.2.3
Jinja2==3.1.6
jiter==0.12.0
jiwer==4.0.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.79
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langsmith==0.4.42
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.1
modelscope==1.31.0
more-itertools==10.8.0
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.4
nvidia-cublas-cu12==12.9.0.13
nvidia-cuda-cccl-cu12==12.9.27
nvidia-cuda-cupti-cu12==12.9.19
nvidia-cuda-nvrtc-cu12==12.9.41
nvidia-cuda-runtime-cu12==12.9.37
nvidia-cudnn-cu12==9.9.0.52
nvidia-cufft-cu12==11.4.0.6
nvidia-cufile-cu12==1.14.0.30
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.4.40
nvidia-cusparse-cu12==12.5.9.5
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.9.41
nvidia-nvtx-cu12==12.9.19
openai==2.8.0
opencv-contrib-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.11.4
packaging==25.0
paddleocr==3.3.2
paddlepaddle-gpu==3.2.2
paddlex==3.3.9
pandas==2.3.3
pdf2image==1.17.0
pillow==12.0.0
premailer==3.10.0
prettytable==3.16.0
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.3
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycryptodome==3.23.0
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pynvml==13.0.1
pypdfium2==5.0.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.14.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentencepiece==0.2.1
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
tqdm==4.67.1
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
wcwidth==0.2.14
websocket-client==1.9.0
yarl==1.22.0
zstandard==0.25.0
